ansible: extend fuzzers (#7811)
* ansible: rename fuzzers

* ansible: add encryption fuzzer
