+++ projects/ansible/build.sh
@@ -15,12 +15,18 @@
 #
 ################################################################################
 
-# Build and install project (using current CFLAGS, CXXFLAGS).
-pip3 install --upgrade pip
-pip3 install -r ./requirements.txt
 pip3 install .
 
-# Build fuzzers in $OUT.
+cp /usr/lib/x86_64-linux-gnu/libcrypt.so.1.1.0 /out/libcrypt.so
 cd $SRC
-compile_python_fuzzer parse_fuzz.py --add-data ansible/lib/ansible/config:ansible/config
-compile_python_fuzzer task_fuzz.py --add-data ansible/lib/ansible/config:ansible/config
+
+# Build parse and task fuzzers
+compile_python_fuzzer fuzz_parse.py --add-data ansible/lib/ansible/config:ansible/config
+compile_python_fuzzer fuzz_task.py --add-data ansible/lib/ansible/config:ansible/config
+
+# Build fuzz_encrypt with a specific wrapper only in non-coverage
+if [ "$SANITIZER" != "coverage" ]; then
+  compile_python_fuzzer fuzz_encrypt.py --add-data ansible/lib/ansible/config:ansible/config
+  cp $SRC/fuzz_encrypt.sh $OUT/fuzz_encrypt
+  chmod +x $OUT/fuzz_encrypt
+fi
